forward rendering

+++ cuda_rasterizer/rasterizer_impl.h
@@ -14,6 +14,7 @@
 #include <iostream>
 #include <vector>
 #include "rasterizer.h"
+#include <cuda_fp16.h>
 #include <cuda_runtime_api.h>
 
 namespace CudaRasterizer
@@ -55,8 +56,8 @@ namespace CudaRasterizer
 	struct BinningState
 	{
 		size_t sorting_size;
-		uint64_t* point_list_keys_unsorted;
-		uint64_t* point_list_keys;
+		uint32_t* point_list_keys_unsorted;
+		uint32_t* point_list_keys;
 		uint32_t* point_list_unsorted;
 		uint32_t* point_list;
 		char* list_sorting_space;
